+++ src/Cmark/File/LogoManager.h
