+++ src/Cmark/File/Resolver/EXIFResolver.h
