+++ src/Cmark/File/Resolver/EasyExifResolver.cpp
