+++ src/Cmark/File/Resolver/EasyExifResolver.h
