+++ src/Cmark/Scene/CScene.h
@@ -7,7 +7,7 @@
 #include "Base/Type.h"
 #include "SceneLayoutSettings.h"
 
-#include "sources/LogoManager.h"
+#include "File/LogoManager.h"
 
 namespace CM
 {
@@ -151,7 +151,7 @@ namespace CM
 
 #ifdef TOOLWIDGET
         QGraphicsWidget* m_ToolWidget;    ///< 悬浮窗口
-#endif 
+#endif
 
         PreViewImageItem* m_ShowImageItem{}; ///< 预览加载的图片Item对象
         QGraphicsPixmapItem* m_LogoItem{}; ///< 加载的logo Item对象
