+++ src/Cmark/Window/ImagePropertyDockWidget.cpp
