+++ src/Cmark/Window/ImagePropertyDockWidget.h
