+++ src/Cmark/Window/LeftDockWidget.cpp
